+++ src/api/leagues.jsx
@@ -4,7 +4,7 @@ import {validateResponse} from '../helpers/response';
 
 const fetchAllLeagues = async accessToken => {
   try {
-    const responseIncludeFields = 'id,name,sport_kind,background_url';
+    const responseIncludeFields = 'id,name,sport_kind,slug';
     const response = await fetch(
       `${API_HOST}/api/v1/leagues?response_include_fields=${responseIncludeFields}&api_access_token=${accessToken}`,
       {