+++ src/assets/images/index.jsx
@@ -1,0 +1,9 @@
+export const LeaguesAssets = {
+  bundesliga: require('./leagues/bundesliga.png'),
+  champions_league: require('./leagues/champions_league.png'),
+  epl: require('./leagues/epl.png'),
+  la_liga: require('./leagues/la_liga.png'),
+  nba: require('./leagues/nba.png'),
+  rpl: require('./leagues/rpl.png'),
+  seria_a: require('./leagues/seria_a.png'),
+};